Many Functions Generalized & Fewer Complaint Form Pages
While debugging a long-known issue with something, i opted to upgrade
to the conditional logic that nodes can perform live on the page
(js/jquery). Now each response to a node can slide down any specific
child node it desires (instead of all-kids-or-nothing).

Also, i finally figured out the lingering issue with the Allegations,
which was half smoke in mirrors. This empowered the ability to create
multiple, different allegations on the same page. Which is nice.

With these engine upgrades, and in line with Evan's advice — not to be
shy of longer pages — we've just gone from 86 possible pages down to
50!-D The Gold section lost the most with maybe 20 pages, and we are
generally much closer to one-page-per-menu-sub-item.
https://openpolice.org/tree/complaint

Especially until i'm able to dedicate time to optimizing load-time in
between complaint form pages, this should help significantly reduce the
total wait times during the process.