Gold Level Upgrade
No longer really repeating any questions in the Gold section! We are
only asking details once for each type of event, instead of once
per-person-who-the-event-happened-for.

[For complaints with multiple victims...] The questions asking which
specific Victims had Stops, Searches, Force, and Arrest has been moved
from the Gold to the Silver. Previously, the Silver version asked
'which victim(s)' as a followup question to most Allegations. Now, I've
ditched the links between Victims and Allegations for Stops, Searches,
Force, and Arrests (Events). But now we're linking Victims to Events in
Silver instead of in Gold — quite a valuable data boost, imho — without
adding to the question count or complexity of the Silver experience.